重新整理了entity，只有cart类有错误。

+++ src/main/java/com/hongqiang/shop/modules/product/web/member/ProductNotifyController.java
@@ -12,6 +12,7 @@ import com.hongqiang.shop.common.web.BaseController;
 import com.hongqiang.shop.modules.entity.Member;
 import com.hongqiang.shop.modules.entity.ProductNotify;
 import com.hongqiang.shop.modules.product.service.ProductNotifyService;
+import com.hongqiang.shop.modules.user.service.MemberService;
 
 @Controller("shopMemberProductNotifyController")
 @RequestMapping({"/member/product_notify"})
@@ -22,15 +23,15 @@ public class ProductNotifyController extends BaseController
   @Autowired
   ProductNotifyService productNotifyService;
 
-//  @Autowired
-//  MemberService memberService;
+  @Autowired
+  MemberService memberService;
 
   @RequestMapping(value={"/list"}, method={org.springframework.web.bind.annotation.RequestMethod.GET})
   public String list(Integer pageNumber, Model model)
   {
-//    Member localMember = this.memberService.getCurrent();
+    Member localMember = this.memberService.getCurrent();
     Pageable localPageable = new Pageable(pageNumber, Integer.valueOf(PAGE_SIZE));
-//    model.addAttribute("page", this.productNotifyService.findPage(localMember, null, null, null, localPageable));
+    model.addAttribute("page", this.productNotifyService.findPage(localMember, null, null, null, localPageable));
     return "/shop/member/product_notify/list";
   }
 
@@ -42,25 +43,10 @@ public class ProductNotifyController extends BaseController
     ProductNotify localProductNotify = (ProductNotify)this.productNotifyService.find(id);
     if (localProductNotify == null)
       return ADMIN_ERROR;
-//    Member localMember = this.memberService.getCurrent();
-//    if (!localMember.getProductNotifies().contains(localProductNotify))
-//      return ADMIN_ERROR;
+    Member localMember = this.memberService.getCurrent();
+    if (!localMember.getProductNotifies().contains(localProductNotify))
+      return ADMIN_ERROR;
     this.productNotifyService.delete(localProductNotify);
     return ADMIN_SUCCESS;
   }
-
-//  @RequestMapping({"delete"})
-//  @ResponseBody
-//  public Message delete(Long id)
-//  {
-//    ProductNotify localProductNotify = (ProductNotify)this.productNotifyService.find(id);
-//    if (localProductNotify == null)
-//      return IIIllIll;
-//    Member localMember = this.memberService.getCurrent();
-//    if (!localMember.getProductNotifies().contains(localProductNotify))
-//      return IIIllIll;
-//    this.productNotifyService.delete(localProductNotify);
-//    return Message.success("admin.message.success",null);
-//  }
-
 }
